fitness_trainer/requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.3
numpy==1.24.3
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Pillow==10.0.1
scipy==1.11.3
tensorflow==2.13.0
scikit-learn==1.3.1
matplotlib==3.8.1

